[PDW-67] fix: 회의실 관리 제목바 수정
